Add getTexLevelParameter and corresponding features into the spec
getTexLevelParameter is not listed in ARB_texture_multisample. However,
it has been added in OpenGL 3.1 core profile. ARB_texture_multisample
is written against OpenGL 3.1. So it is oK to expose it at GL side.
In addition, getTexLevelParameter is in OpenGL ES 3.1. So it is OK to
expose it at GLES side too.